334: Add license_last_16_weeks and community_last_16_weeks
Add both these fields to the database and all appropriate workload object
instaniations. Update the integration tests to ensure these fields are retrieved
from the database in the correct order.

This branch requires the branch in the worker to be merged before it will work
as it depends on changes in the context mapper.