Feature/wmt 678/contracted hours (#55)
* 678: Add contracted hours to workload points calculation

Remove working_hours table, and include contracted hours in the workload_owner
table. Update unit and integration tests to account for working_hours
table no longer existing.

Update contracted_hours and reductions to be float, as opposed to int

* 678: Update formatting

* 678: Make contracted_hours in WPC notNullable and removed unneeded variables

* 678: Add script to delete views (#58)

* 678: Add script to delete views

This needs to be run before a rollback so the base tables can be deleted.

Run using: `node delete-views.js`

* Renamed variables

* Feature/wmt 678/add seed directory (#57)

* 678: Move seed data files to

* 678: Add views seed directory

Database object which don't require the write once functionality provided by
migrations should be handled using seeds. This means that each files is ran on
each deployment.

This gives us the flexibility of being able to change database views without
needing to create a new file.

These are numbered sequentially and can be run using the following command:

`yarn seed-views`

* 678: Remove unnecessary fields from capacity views

Now that the worker is calaculating the total points we need to reflect that in
the seed data to ensure the capacity graph is displayed correctly.

* 678: Renamed files and variables

* 678: Add contracted hours to inserts and tests

Also, update float data types to decimal