1041: Update worker to allow for UPDATE calculate workload points task
If a workload points calculation is triggered from the web app, the WPC should be recalculated
and the relevant rows updated in the DB, as opposed to new ones being inserted. The workload report
should not be affected.

New field 'operationType' added to the additional data on the task to check whether to insert or update