Feature/wmt 414/display arms assessments workload migrations (#115)
* 414: Add arms table to staging schema

* 414: Add required columns to the workload table

Add arms_license_cases and arms_community_cases which will store counts of the
number of cases of that type stored in the workload. These numbers will then be
used by the points caluclation worker to adjust the total points for a workload
owner.

* 414: Add ARMS case counts to the seed data

* 414: Update yarn lock file for probation rules

Make sure we're usin the correct version of probabtion rules i.e. 0.1.12

* 414: Fix typo in migration file name