Create one knex instance per user instead of one per file.
Before this change the application created a new knex instance, and therefore a
new connection pool, each time it was reference from a file. That means we
couldn't control the size of the pool for the application.

We need two pools, one for each user, that is one for the application schema
user and one for the staging schema user.