Feature/wmt 414/display arms assessments (#118)
* 414: Add arms table to staging schema

* 414: Move points calculations to the probation rules

The probation rules `calculateWorkloadPoints` should encompass PAROMS and SDRs.

* 414: Add required fields to get app workloads query

* 414: Join on om_key and team_key when querying staging workloads

In order to ensure workloads are imported for workload owners i.e. the unique
combination of offender managers and teams. The workloads should ensure that
both of these fields match when joining the staging tables.

* 414: Add arms data service

In order to make sure the correct values are being retrieved from the staging
schema the ARMS properties needed to be retrieved.

* 414: Add team_code to test data

* 414: Move calculateWeighting method from worker into probation rules

* 414: Fix indenting

* 414: Capitalise keywords in SQL

* 414: Add arms total to workload points calculations

Users want to be able to see the number of ARMS assessments which make up a
capacity. This is displayed in the capacity breakdown table on each level of the
frontend.

This requires changes in the workload points calculations worker, the capacity
breakdown views and a new database migration.